refactor: split the game context
